#81 make avatar icon responsive
divided appheader into several components so that they can be used in
mobilemenu and appheader
