+++ src/components/Header/mobile.css
@@ -82,6 +82,10 @@
     width: 100%;
   }
 
+  .login-buttons {
+    width: 10rem;
+  }
+
   .input{
     width: 100%;
     height: 40px;
